pyproject.toml
[tool.poetry]
name = "player-stats"
version = "0.1.0"
description = ""
authors = ["wbglaeser <ben.glaeser@soundcloud.com>"]
readme = "README.md"
packages = [{include = "player_stats"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "2.28.2"
beautifulsoup4 = "^4.11.2"
gspread = "^5.7.2"
crcmod = "^1.7"
google-cloud-kms = "^2.15.0"
google-cloud-storage = "^2.7.0"
pandas = "^1.5.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
